메소드 추출을 통해 수정함
